fix: move notifications to correct folder
