+++ web/src/components/pages/PlaygroundPage/PlaygroundPage.tsx
@@ -7,7 +7,7 @@ import { dispatchLoadSnippet } from '~/store/workspace'
 import { Header } from '~/components/layout/Header'
 import { ConnectedWorkspace } from '~/components/features/workspace/Workspace'
 import { InspectorPanel } from '~/components/features/inspector/InspectorPanel/InspectorPanel'
-import { NotificationHost } from '~/components/modals/Notification'
+import { NotificationHost } from '~/components/layout/Notification'
 import { Layout } from '~/components/layout/Layout/Layout'
 import { ConnectedStatusBar } from '~/components/layout/StatusBar'
 import { computeSizePercentage } from './utils'
